+++ sudo_plugin/src/macros.rs
@@ -20,40 +20,38 @@
 /// # Example
 ///
 /// ```rust
+/// # mod necessary_for_super_type_lookup_to_work {
 /// use sudo_plugin::*;
 /// use sudo_plugin::errors::*;
 /// use std::io::Write;
 ///
-/// sudo_io_plugin! {
-///     example : Example {
-///         close:      close,
-///         log_stdout: log_stdout,
-///     }
-/// }
+/// sudo_io_plugin! { example : Example }
 ///
 /// struct Example {
-///     plugin: &'static sudo_plugin::Plugin
+///     env: &'static sudo_plugin::IoEnv
 /// }
 ///
-/// impl Example {
-///     fn open(plugin: &'static sudo_plugin::Plugin) -> Result<Self> {
-///         plugin.stdout().write(b"example sudo plugin initialized");
+/// impl IoPlugin for Example {
+///     const NAME:    &'static str = "example";
+///     const VERSION: &'static str = env!("CARGO_PKG_VERSION");
 ///
-///         Ok(Example { plugin })
+///     fn open(env: &'static sudo_plugin::IoEnv) -> Result<Self> {
+///         writeln!(env.stdout(), "example sudo plugin initialized");
+///
+///         Ok(Example { env })
 ///     }
 ///
-///     fn close(&mut self, _: i32, _: i32) {
-///         self.plugin.stdout().write(b"example sudo plugin exited");
+///     fn close(self, _: i32, _: i32) {
+///         writeln!(self.env.stdout(), "example sudo plugin exited");
 ///     }
 ///
 ///     fn log_stdout(&mut self, _: &[u8]) -> Result<()> {
-///         self.plugin.stdout().write(
-///             b"example sudo plugin received output on stdout"
-///         );
+///         writeln!(self.env.stdout(), "example sudo plugin received output on stdout");
 ///
 ///         Ok(())
 ///     }
 /// }
+/// # }
 /// ```
 ///
 /// The generated plugin will have the entry point `example`, so to
@@ -66,223 +64,40 @@
 /// ```
 #[macro_export]
 macro_rules! sudo_io_plugin {
-    ( $name:ident : $ty:ty { $($cb:ident : $fn:ident),* $(,)* } ) => {
-        use ::sudo_plugin::errors::AsSudoPluginRetval;
+    ( $name:ident : $ty:ty ) => {
+        mod $name {
+            use super::*;
 
-        static mut PLUGIN:   Option<::sudo_plugin::Plugin> = None;
-        static mut INSTANCE: Option<$ty>                   = None;
+            // TODO: end use of static mut
+            static mut SUDO_IO_ENV:    Option<$crate::IoEnv> = None;
+            static mut SUDO_IO_PLUGIN: Option<$ty>           = None;
 
-        #[no_mangle]
+            pub struct State { }
+
+            impl $crate::IoState<$ty> for State {
+                unsafe fn io_env()    -> &'static mut Option<$crate::IoEnv> { &mut SUDO_IO_ENV }
+                unsafe fn io_plugin() -> &'static mut Option<$ty>           { &mut SUDO_IO_PLUGIN }
+            }
+        }
+
         #[allow(non_upper_case_globals)]
         #[allow(missing_docs)]
-        pub static $name: ::sudo_plugin::sys::io_plugin = {
-            ::sudo_plugin::sys::io_plugin {
-                // construct the plugin using any callbacks specified
-                $( $cb: sudo_io_fn!($cb, $name, PLUGIN, INSTANCE, $fn) ),*,
+        #[no_mangle]
+        pub static $name: $crate::sys::io_plugin = $crate::sys::io_plugin {
+            type_:   $crate::sys::SUDO_IO_PLUGIN,
+            version: $crate::sys::SUDO_API_VERSION,
 
-                // and for anything not specified, use the defaults
-                .. ::sudo_plugin::sys::io_plugin {
-                    type_:            ::sudo_plugin::sys::SUDO_IO_PLUGIN,
-                    version:          ::sudo_plugin::sys::SUDO_API_VERSION,
-                    open:             Some(open),
-                    close:            Some(close),
-                    show_version:     Some(show_version),
-                    .. ::sudo_plugin::sys::IO_PLUGIN_EMPTY
-                }
-            }
+            open:         Some($crate::core::open::<$ty, $name::State>),
+            close:        Some($crate::core::close::<$ty, $name::State>),
+            show_version: Some($crate::core::show_version::<$ty, $name::State>),
+
+            log_ttyin:  Some($crate::core::log_ttyin ::<$ty, $name::State>),
+            log_ttyout: Some($crate::core::log_ttyout::<$ty, $name::State>),
+            log_stdin:  Some($crate::core::log_stdin ::<$ty, $name::State>),
+            log_stdout: Some($crate::core::log_stdout::<$ty, $name::State>),
+            log_stderr: Some($crate::core::log_stderr::<$ty, $name::State>),
+
+            .. ::sudo_plugin::sys::IO_PLUGIN_EMPTY
         };
-
-        unsafe extern "C" fn open(
-            version:            ::libc::c_uint,
-            conversation:       ::sudo_plugin::sys::sudo_conv_t,
-            plugin_printf:      ::sudo_plugin::sys::sudo_printf_t,
-            settings_ptr:       *const *mut ::libc::c_char,
-            user_info_ptr:      *const *mut ::libc::c_char,
-            command_info_ptr:   *const *mut ::libc::c_char,
-            argc:               ::libc::c_int,
-            argv:               *const *mut ::libc::c_char,
-            user_env_ptr:       *const *mut ::libc::c_char,
-            plugin_options_ptr: *const *mut ::libc::c_char,
-        ) -> ::libc::c_int {
-            let (mut stdout, mut stderr) = ::sudo_plugin::plugin::PrintFacility::new(
-                Some(stringify!($name)), plugin_printf
-            );
-
-            let plugin = ::sudo_plugin::Plugin::new(
-                stringify!($name).into(),
-                option_env!("CARGO_PKG_VERSION").map(Into::into),
-                version,
-                argc, argv,
-                settings_ptr,
-                user_info_ptr,
-                command_info_ptr,
-                user_env_ptr,
-                plugin_options_ptr,
-
-                stdout,
-                stderr.clone(), // we need stderr ourselves if `open` fails
-                conversation,
-            );
-
-            match plugin {
-                Ok(p)  => PLUGIN = Some(p),
-                Err(e) => {
-                    let _ = stderr.write_error(&e);
-                    return e.as_sudo_io_plugin_open_retval();
-                },
-            };
-
-            // unwrap should be panic-safe here, since we just assigned
-            // a value to $plugin
-            let plugin = PLUGIN.as_ref().unwrap();
-
-            // if the command is empty, to the best of my knowledge
-            // we're being called with `-V` to report our version; in
-            // this case there's no reason to fully invoke the plugin
-            // through its `open` function
-            if plugin.command_info.command == ::std::path::PathBuf::default() {
-                return ::sudo_plugin::sys::SUDO_PLUGIN_OPEN_SUCCESS;
-            }
-
-            // call the plugin's `open` function
-            match <$ty>::open(plugin) {
-                Ok(i)  => INSTANCE = Some(i),
-                Err(e) => {
-                    let e: ::sudo_plugin::errors::Error = e.into();
-                    let _ = stderr.write_error(&e);
-                    return e.as_sudo_io_plugin_open_retval();
-                },
-            }
-
-            ::sudo_plugin::sys::SUDO_PLUGIN_OPEN_SUCCESS
-        }
-
-        unsafe extern "C" fn close(
-            _exit_status: ::libc::c_int,
-            _error:       ::libc::c_int,
-        ) {
-            // force the instance to be dropped
-            let _ = INSTANCE.take();
-        }
-
-        unsafe extern "C" fn show_version(
-            _verbose: ::libc::c_int,
-        ) -> ::libc::c_int {
-            if let Some(plugin) = PLUGIN.as_ref() {
-                // disable the write_literal lint since it has a known
-                // bug that fires when you use a macro that expands to
-                // a literal (e.g., `stringify!`)
-                #[cfg_attr(feature="cargo-clippy", allow(clippy::write_literal))]
-                let _ = writeln!(
-                    plugin.stdout(),
-                    "{} I/O plugin version {}",
-                    plugin.plugin_name,
-                    plugin.plugin_version.as_deref().unwrap_or("<unknown>"),
-                );
-            }
-
-            0
-        }
     }
-}
-
-/// Internal macro used by `sudo_io_plugin` that  generates the actual
-/// callback implementations for I/O plugins.
-#[macro_export]
-macro_rules! sudo_io_fn {
-    ( close , $name:tt , $plugin:expr , $instance:expr , $fn:ident ) => {{
-        unsafe extern "C" fn close(
-            exit_status: ::libc::c_int,
-            error:       ::libc::c_int,
-        ) {
-            if let Some(mut i) = $instance.take() {
-                i.$fn(exit_status as _, error as _);
-            }
-        }
-
-        Some(close)
-    }};
-
-    ( log_ttyin , $name:tt, $plugin:expr , $instance:expr , $fn:ident ) => {
-        sudo_io_fn!(log, log_ttyin, $name, $plugin, $instance, $fn)
-    };
-
-    ( log_ttyout , $name:tt, $plugin:expr , $instance:expr , $fn:ident ) => {
-        sudo_io_fn!(log, log_ttyout, $name, $plugin, $instance, $fn)
-    };
-
-    ( log_stdin , $name:tt, $plugin:expr , $instance:expr , $fn:ident ) => {
-        sudo_io_fn!(log, log_stdin, $name, $plugin, $instance, $fn)
-    };
-
-    ( log_stdout , $name:tt, $plugin:expr , $instance:expr , $fn:ident ) => {
-        sudo_io_fn!(log, log_stdout, $name, $plugin, $instance, $fn)
-    };
-
-    ( log_stderr , $name:tt, $plugin:expr , $instance:expr , $fn:ident ) => {
-        sudo_io_fn!(log, log_stderr, $name, $plugin, $instance, $fn)
-    };
-
-    (
-        log ,
-        $log_fn:ident ,
-        $name:tt ,
-        $plugin:expr ,
-        $instance:expr ,
-        $fn:ident
-    ) => {{
-        unsafe extern "C" fn $log_fn(
-            buf: *const ::libc::c_char,
-            len:        ::libc::c_uint,
-        ) -> ::libc::c_int {
-            let slice = ::std::slice::from_raw_parts(
-                buf as *const _,
-                len as _,
-            );
-
-            let result : ::std::result::Result<(), ::sudo_plugin::errors::Error> = $instance
-                .as_mut()
-                .map_or_else(
-                  || Err(::sudo_plugin::errors::ErrorKind::Uninitialized.into()),
-                  |i| i.$fn(slice).map_err(|e| e.into()),
-                );
-
-            // if there was an error (and we can unwrap the plugin),
-            // write it out
-            if let (Some(p), Err(e)) = ($plugin.as_ref(), result.as_ref()) {
-                let _ = p.stderr().write_error(&e);
-            }
-
-            result.as_sudo_io_plugin_log_retval()
-        }
-
-        Some($log_fn)
-    }};
-
-    ( change_winsize , $name:tt , $plugin:expr , $instance:expr , $fn:ident ) => {{
-        unsafe extern "C" fn change_winsize(
-            lines: ::libc::c_uint,
-            cols:  ::libc::c_uint,
-        ) {
-            let result : ::std::result::Result<(), ::sudo_plugin::errors::Error> = $instance
-                .as_mut()
-                .map_or_else(
-                  || Err(::sudo_plugin::errors::ErrorKind::Uninitialized.into()),
-                  |i| i.$fn(slice).map_err(|e| e.into()),
-                );
-
-            // if there was an error (and we can unwrap the plugin),
-            // write it out
-            if let (Some(p), Err(e)) = ($plugin.as_ref(), result.as_ref()) {
-                let _ = p.stderr().write_error(
-                    p.plugin_name.as_bytes(),
-                    e,
-                );
-            }
-
-            result.as_sudo_io_plugin_log_retval()
-        }
-
-        Some(change_winsize)
-    }};
 }